testes de adição, edição e remoção de cupons de admin e restaurante funcionando
