+++ server/admin-coupons.json
@@ -1,1 +1,1 @@
-[{"id":"2835c648","name":"SEGUNDACOMPRA","adm":true,"minValue":15,"product":"Nenhum","discount":0.5,"status":"Ativo"}]
+[{"id":"fd4f7191","name":"SEGUNDACOMPRA","adm":true,"minValue":15,"product":"Nenhum","discount":0.5,"status":"Ativo"}]
