+++ server/restaurants-coupons.json
@@ -1,108 +1,1 @@
-[
-    {
-        "name": "Mequi",
-        "products": [
-            {
-                "name": "Big Méqui",
-                "price": 15
-            },
-            {
-                "name": "Cheddar Méquimelt",
-                "price": 12
-            },
-            {
-                "name": "Méqui Duplo",
-                "price": 12.5
-            },
-            {
-                "name": "Méquishake",
-                "price": 10
-            },
-            {
-                "name": "Méqui Chicken",
-                "price": 10
-            },
-            {
-                "name": "Méqui Tasty",
-                "price": 18
-            },
-            {
-                "name": "Méqui Fritas",
-                "price": 9
-            }
-        ],
-        "coupons": [
-            {
-                "id": "22b9de6a",
-                "name": "A",
-                "adm": false,
-                "minValue": 10,
-                "product": "Méquishake",
-                "discount": 0.4,
-                "status": "Inativo"
-            },
-            {
-                "id": "a4f3f3de",
-                "name": "ASASAS",
-                "adm": false,
-                "minValue": 0.14,
-                "product": "Méquishake",
-                "discount": 0.5,
-                "status": "Ativo"
-            },
-            {
-                "id": "da488bca",
-                "name": "2BIGMEQUIPOR1",
-                "adm": false,
-                "minValue": 15,
-                "product": "Big Méqui",
-                "discount": 0.5,
-                "status": "Ativo"
-            }
-        ]
-    },
-    {
-        "name": "BK",
-        "products": [
-            {
-                "name": "Big BK",
-                "price": 14.5
-            },
-            {
-                "name": "Whopper Duplo",
-                "price": 16
-            },
-            {
-                "name": "Whopper",
-                "price": 12
-            },
-            {
-                "name": "BKshake",
-                "price": 12.5
-            },
-            {
-                "name": "BK Chicken",
-                "price": 10.5
-            },
-            {
-                "name": "Whopper Furioso",
-                "price": 17
-            },
-            {
-                "name": "BK Fritas",
-                "price": 8.5
-            }
-        ],
-        "coupons": [
-            {
-                "id": "69d5833c",
-                "name": "60OFFWHOPPER",
-                "adm": false,
-                "minValue": 32,
-                "product": "Whopper",
-                "discount": 0.2,
-                "status": "Inativo"
-            }
-        ]
-    }
-]
+[{"name":"Mequi","products":[{"name":"Big Méqui","price":15},{"name":"Cheddar Méquimelt","price":12},{"name":"Méqui Duplo","price":12.5},{"name":"Méquishake","price":10},{"name":"Méqui Chicken","price":10},{"name":"Méqui Tasty","price":18},{"name":"Méqui Fritas","price":9}],"coupons":[{"id":"22b9de6a","name":"A","adm":false,"minValue":10,"product":"Méquishake","discount":0.4,"status":"Inativo"},{"id":"a4f3f3de","name":"ASASAS","adm":false,"minValue":0.14,"product":"Méquishake","discount":0.5,"status":"Ativo"},{"id":"ddd61822","name":"2BIGMEQUIPOR1","adm":false,"minValue":15,"product":"Big Méqui","discount":0.5,"status":"Ativo"}]},{"name":"BK","products":[{"name":"Big BK","price":14.5},{"name":"Whopper Duplo","price":16},{"name":"Whopper","price":12},{"name":"BKshake","price":12.5},{"name":"BK Chicken","price":10.5},{"name":"Whopper Furioso","price":17},{"name":"BK Fritas","price":8.5}],"coupons":[{"id":"69d5833c","name":"60OFFWHOPPER","adm":false,"minValue":32,"product":"Whopper","discount":0.2,"status":"Inativo"}]}]
